backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
email-validator==2.1.0.post1
langchain==0.2.16
langchain-openai==0.1.23
# pypdf==5.1.0  # Removed - using only OCR
# pdfminer.six==20231228  # Removed - using only OCR
pytesseract==0.3.13
pdf2image==1.17.0

# RAG Dependencies
qdrant-client>=1.9.0,<2.0.0
openai>=1.54.0,<2.0.0
tiktoken>=0.7,<1.0
sentence-transformers==2.2.2
numpy==1.24.3
google-generativeai==0.3.2

# Autonomous Agents with LangGraph
langgraph>=0.2.16
langchain-core>=0.2.27,<0.3
langchain-community>=0.2.16
langchain-qdrant>=0.1.0
langchain-text-splitters>=0.2.4

